requirements.txt
celery
redis
tqdm
Flask
python-dotenv
